MainActivity内のFragment処理を分離
